+++ maplibrary/layers/encounterslayer.h
@@ -18,7 +18,7 @@ EncounterType tag_invoke(js::value_to_tag<EncounterType>, js::value const& jv);
 class EncountersLayer : public Layer<EncounterType>
 {
 public:
-    EncountersLayer(int nCol, int nRow);
+    EncountersLayer(size_t nCol, size_t nRow);
 
     friend void tag_invoke(js::value_from_tag, js::value& jv, std::unique_ptr<EncountersLayer> const& o);
     friend std::unique_ptr<EncountersLayer> tag_invoke(js::value_to_tag<std::unique_ptr<EncountersLayer>>,
