+++ maplibrary/layers/eventlayer.cpp
@@ -1,6 +1,6 @@
 #include "eventlayer.h"
 
-EventLayer::EventLayer(int nCol, int nRow) : Layer(nCol, nRow)
+EventLayer::EventLayer(size_t nCol, size_t nRow) : Layer(nCol, nRow)
 {
 }
 
@@ -15,7 +15,7 @@ std::unique_ptr<EventLayer> tag_invoke(js::value_to_tag<std::unique_ptr<EventLay
 {
     js::object const&           obj = jv.as_object();
     std::unique_ptr<EventLayer> o =
-        std::make_unique<EventLayer>(js::value_to<int>(obj.at("nCol")), js::value_to<int>(obj.at("nRow")));
+        std::make_unique<EventLayer>(js::value_to<size_t>(obj.at("nCol")), js::value_to<size_t>(obj.at("nRow")));
     o->elements = js::value_to<std::vector<std::vector<Event::EventPtr>>>(obj.at("elements"));
     return o;
 }
