+++ maplibrary/layers/eventlayer.h
@@ -9,7 +9,7 @@
 class EventLayer : public Layer<Event::EventPtr>
 {
 public:
-    EventLayer(int nCol, int nRow);
+    EventLayer(size_t nCol, size_t nRow);
 
     friend void tag_invoke(js::value_from_tag, js::value& jv, std::unique_ptr<EventLayer> const& o);
     friend std::unique_ptr<EventLayer> tag_invoke(js::value_to_tag<std::unique_ptr<EventLayer>>, js::value const& jv);
